service/api/requirements.txt
Flask
PyYaml==4.2b1